chore: useDectectClose훅 사용하도록 변경
