+++ src/packages/basic-ui/src/Dropdown/hooks/useDropdownState.ts
@@ -4,21 +4,16 @@ import type { DropdownOptionType } from "../index";
 interface UseDropdownStateProps {
   options: DropdownOptionType[];
   value?: string | number;
-  disabled: boolean;
   onChange?: (value: string | number, option: DropdownOptionType) => void;
-  onOpen?: () => void;
   onClose?: () => void;
 }
 
 export function useDropdownState({
   options,
   value,
-  disabled,
   onChange,
-  onOpen,
   onClose,
 }: UseDropdownStateProps) {
-  const [isOpen, setIsOpen] = useState(false);
   const [selectedOption, setSelectedOption] = useState<
     DropdownOptionType | undefined
   >();
@@ -29,32 +24,11 @@ export function useDropdownState({
     setSelectedOption(option);
   }, [value, options]);
 
-  const openDropdown = useCallback(() => {
-    if (disabled) return;
-    setIsOpen(true);
-    onOpen?.();
-  }, [disabled, onOpen]);
-
-  const closeDropdown = useCallback(() => {
-    setIsOpen(false);
-    onClose?.();
-  }, [onClose]);
-
-  const toggleDropdown = useCallback(() => {
-    if (disabled) return;
-    if (isOpen) {
-      closeDropdown();
-    } else {
-      openDropdown();
-    }
-  }, [disabled, isOpen, closeDropdown, openDropdown]);
-
   const selectOption = useCallback(
     (option: DropdownOptionType) => {
       if (option.disabled) return;
 
       setSelectedOption(option);
-      setIsOpen(false);
       onChange?.(option.value, option);
       onClose?.();
     },
@@ -62,11 +36,7 @@ export function useDropdownState({
   );
 
   return {
-    isOpen,
     selectedOption,
-    openDropdown,
-    closeDropdown,
-    toggleDropdown,
     selectOption,
   };
 }
